add meta and favicon and convert client page to server page
