+++ src/components/header.js
@@ -10,12 +10,13 @@ import {
   getAccessToken,
   removeUserSession,
 } from "@/utils/common";
-import { useRouter } from "next/navigation";
+import { useRouter ,usePathname} from "next/navigation";
 import { API_URL } from "../../constant";
 
 const Header = () => {
   const [isClient, setIsClient] = useState(false);
   const router = useRouter();
+  const pathName = usePathname()
   const userName = getUser();
   const accessToken = getAccessToken();
   const refresh = getToken();
@@ -23,6 +24,8 @@ const Header = () => {
   useEffect(() => {
     setIsClient(true);
   }, []);
+
+  console.log(pathName)
 
   const handleLogout = async () => {
     try {
@@ -52,40 +55,45 @@ const Header = () => {
     router.push("/adminpanel");
   };
 
-  return (
-    <>
-      <div className="header bg-[#1D2E3E] p-4 flex items-center justify-between">
-        <Image
-          className="cursor-pointer"
-          alt="Arcgate"
-          onClick={goBack}
-          width={30}
-          src={Backbutton}
-        />
 
-        <div className="logo w-36 md:w-32 text-center md:text-left md:m-auto">
+  if(pathName !== "/"){
+
+    return (
+      <>
+        <div className="header bg-[#1D2E3E] p-4 flex items-center justify-between">
           <Image
-            src={Logo}
+            className="cursor-pointer"
             alt="Arcgate"
-            onClick={() => {
-              router.push("/adminpanel", { scroll: false });
-            }}
+            onClick={goBack}
+            width={30}
+            src={Backbutton}
           />
+  
+          <div className="logo w-36 md:w-32 text-center md:text-left md:m-auto">
+            <Image
+              src={Logo}
+              alt="Arcgate"
+              onClick={() => {
+                router.push("/adminpanel", { scroll: false });
+              }}
+            />
+          </div>
+          <div className="flex gap-x-2 md:gap-x-4 items-center md:items-start mt-2 md:mt-0">
+            <p className="text-[#ffff] uppercase text-lg md:text-xl">
+              {isClient ? userName : "username"}
+            </p>
+            <Image
+              onClick={handleLogout}
+              src={LogoutImage}
+              alt="LogoutImage"
+              className="w-5 md:w-8 cursor-pointer"
+            />
+          </div>
         </div>
-        <div className="flex gap-x-2 md:gap-x-4 items-center md:items-start mt-2 md:mt-0">
-          <p className="text-[#ffff] uppercase text-lg md:text-xl">
-            {isClient ? userName : "username"}
-          </p>
-          <Image
-            onClick={handleLogout}
-            src={LogoutImage}
-            alt="LogoutImage"
-            className="w-5 md:w-8 cursor-pointer"
-          />
-        </div>
-      </div>
-    </>
-  );
+      </>
+    );
+  }
+
 };
 
 export default Header;
